+++ src/server/migrate/qb.ts
@@ -1,7 +1,7 @@
 import { oxmysql } from '@overextended/oxmysql';
 import { triggerClientCallback } from '../utils';
-import { saveAppearance } from '../appearance';
 import { TAppearance } from '@typings/appearance';
+import { saveAppearance } from '../init';
 
 const delay = (ms: number) => new Promise(res => setTimeout(res, ms));
 
@@ -13,7 +13,8 @@ const migrate = async (src: string) => {
         emitNet('qb-clothes:loadSkin', src, 0, element.model, element.skin);
         await delay(200);
         const response = await triggerClientCallback('bl_appearance:client:getAppearance', src) as TAppearance
-        await saveAppearance(src, element.citizenid, response)
+        const playerSrc = parseInt(src)
+        await saveAppearance(playerSrc, element.citizenid, response)
     }
     console.log('Converted '+ response.length + ' appearances')
 };
